Create the `SectionWrapper` component
I have created the `SectionWrapper` component. As of now, to pass images to be used as backgrounds the image paths have to be relative to `/public/` - this means that the ordinary method of using `import` won't work.

It still needs some work: I may try to incorporate props that will allow for fine control of the positioning of the background image. Personally, I think it will probably be better in the long run to remake the background blobbies that already exist to accommodate already-created components, but the addition of those fine positioning controls will probably still be beneficial.

As of right now it looks TERRIBLE with pre-existing components and background blobbies. Other components will require modification (e.g. I intend to fix `TopSection` to accommodate this new background component).