Fix issues with adding the background graphic and then merge (#22)
* Add background graphic and some small changes to `TopSection`

Added the background graphic file to `public/home`, and made some small changes to `TopSection` to accomodate for the top graphic. Note that this still has a lot of issues right now at small viewport breakpoints that I'm trying to resolve.

* Fix the gap that appears at small breakpoints

* Run the linter on everything

* Update TopSection.jsx

* Update TopSection.jsx

* Create the `SectionWrapper` component

I have created the `SectionWrapper` component. As of now, to pass images to be used as backgrounds the image paths have to be relative to `/public/` - this means that the ordinary method of using `import` won't work.

It still needs some work: I may try to incorporate props that will allow for fine control of the positioning of the background image. Personally, I think it will probably be better in the long run to remake the background blobbies that already exist to accommodate already-created components, but the addition of those fine positioning controls will probably still be beneficial.

As of right now it looks TERRIBLE with pre-existing components and background blobbies. Other components will require modification (e.g. I intend to fix `TopSection` to accommodate this new background component).

* Add in parameters for vertical and horizontal offsets

The parameters `verticalOffset`, `horizontalOffset`, `mobileVerticalOffset` and `mobileHorizontalOffset` have been added, allowing for fine positioning control of the `background-image`

* Update index.jsx

* Update TopSection.jsx

* Update index.jsx

* Prettier fixes